Fix: Don’t abort Amazon load on certain $ amounts
Fix: Properly format reload amount as currency in the console table
output.

Fix: The reload amount was stored internally by the app as a number, so
0.50 was stored internally as 0.5. The stored value was then sent to
Amazon as a URL parameter so that it was preloaded upon page load, but
further code checked whether the number in Amazon matched what it truly
should be, and typed the correct value if not. This resulted in an
incorrect value being entered - but not submitted - as further code
verified the value in the web interface a final time before submitting.
In Amazon.reloadCard() we now pass the number to Amazon properly
formatted as currency from the start. Although the code to type the
value shouldn’t need to run, we again clear before typing just to be
sure. Thanks to @mcgwebdesign for reporting the issue.

Update dependencies to current releases.